Make test output friendlier
Using mocha and chai to check answers resulted in very unfriendly
and verbose output whenever an answer was incorrect. This is useful
for actual tests but not for these exercises, and they happen to be
kind of overwhelming/scary for a beginner

Here, I've made a rudimentary testing "framework" that only has
the bare minimum feedback that I felt was required.

- Created basic-test-framework.js
  - Contains the functions required to start and end a test run
- Updated launch.json to contain:
  - Run current file
  - Check answers in debug mode
- Modified package.json
  - `npm test` now runs the actual tests
  - `npm run check-answers` runs the exercise answer-checker
- Updated workflow file to refer to `npm test`
- Updated the README for these changes
- Renamed test-exercises.js to exercises.test.js to allow
codesandbox.io to autorun them as if they were tests